+++ src/router/index.ts
@@ -50,7 +50,7 @@ const router = createRouter({
   ]
 })
 
-router.beforeEach((to, from) => {
+router.beforeEach((to) => {
   const appUser = useUserStore()
   if (appUser.isUserLoggedIn && to.name == 'authorization') {
     return false
